agent/requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
rich>=13.0.0
ddgs>=6.0.0
